Revert "Merge pull request #3536 from akvo/issue/3533-collaborator-organisation"
This reverts commit b69080570be4ccb13f05a327c3b3e54def95f7a6, reversing
changes made to 7b95ce0c7a14ecaa49b44c622f371805c5fe2884.
