+++ akvo/rsr/models/project.py
@@ -1099,25 +1099,6 @@ class Project(TimestampsMixin, models.Model):
         # organisations.
         return self.ancestor().id == settings.EUTF_ROOT_PROJECT
 
-    def in_nuffic_hierarchy(self):
-        """Check if the project is a part of the Nuffic hierarchy."""
-        # FIXME: Ideally, we shouldn't need such a function and all
-        # functionality should be generic enough to enable/disable for other
-        # organisations.
-        return self.ancestor().id == settings.NUFFIC_ROOT_PROJECT
-
-    def get_hierarchy_organisation(self):
-        """Return the hierarchy organisation if project belongs to one."""
-
-        from akvo.rsr.models import Organisation
-
-        if self.in_eutf_hierarchy():
-            return Organisation.objects.get(id=settings.EUTF_ORG_ID)
-        elif self.in_nuffic_hierarchy():
-            return Organisation.objects.get(id=settings.NUFFIC_ORG_ID)
-        else:
-            return None
-
     def project_dates(self):
         """ Return the project start and end dates, preferably the actuals. If they are not set, use
             the planned values.
