+++ akvo/rsr/tests/base.py
@@ -7,9 +7,7 @@ from django.conf import settings
 from django.contrib.auth.models import Group
 from django.test import TestCase, Client
 
-from akvo.rsr.models import (
-    User, Employment, Organisation, Project, RelatedProject, Partnership, PublishingStatus
-)
+from akvo.rsr.models import User, Employment
 from akvo.utils import check_auth_groups
 
 
@@ -39,37 +37,6 @@ class BaseTestCase(TestCase):
             user.set_password(password)
             user.save()
         return user
-
-    @staticmethod
-    def create_organisation(name, enable_restrictions=False):
-        """Create an organisation with the given name."""
-        org = Organisation.objects.create(
-            name=name, long_name=name, enable_restrictions=enable_restrictions
-        )
-        return org
-
-    @staticmethod
-    def create_project(title, published=True, public=True):
-        """Create an project with the given title."""
-        project = Project.objects.create(title=title, is_public=public)
-        status = (
-            PublishingStatus.STATUS_PUBLISHED if published else PublishingStatus.STATUS_UNPUBLISHED
-        )
-        project.publishingstatus.status = status
-        project.publishingstatus.save(update_fields=['status'])
-        return project
-
-    @staticmethod
-    def make_parent(parent, project):
-        RelatedProject.objects.create(
-            project=parent,
-            related_project=project,
-            relation=RelatedProject.PROJECT_RELATION_CHILD
-        )
-
-    @staticmethod
-    def make_partner(project, org):
-        Partnership.objects.create(project=project, organisation=org)
 
     @staticmethod
     def make_org_admin(user, org):
