+++ akvo/settings/30-rsr.conf
@@ -105,9 +105,6 @@ NARRATIVE_REPORTS_BETA_ORGS = [3855,]
 EUTF_ROOT_PROJECT = 4401
 EUTF_ORG_ID = 3394
 
-NUFFIC_ROOT_PROJECT = 7350
-NUFFIC_ORG_ID = 4531
-
 # TODO: if more "globals" are added here it's probably time to create a model to hold them
 # Also the report server doesn't have access to these values so they need to be duplicated :-(
 SINGLE_PERIOD_INDICATORS = {
